+++ ui/src/clockface/index.ts
@@ -1,4 +1,5 @@
 // Import Components
+import Alert from './components/alerts/Alert'
 import AutoInput from './components/auto_input/AutoInput'
 import ConfirmationButton from './components/confirmation_button/ConfirmationButton'
 import Dropdown, {DropdownMode} from './components/dropdowns/Dropdown'
@@ -54,6 +55,7 @@ import {
 
 // Fire de lazer
 export {
+  Alert,
   Alignment,
   AutoComplete,
   AutoInput,